Changes in javascript clock file
