+++ app/javascript/custom/clock.js
@@ -1,0 +1,28 @@
+document.addEventListener("DOMContentLoaded", function (event) {
+  currentTime();
+});
+
+function currentTime() {
+  if (document.getElementById("clock")) {
+    console.log(clock)
+    let date = new Date();
+    let hh = date.getHours();
+    let mm = date.getMinutes();
+    let ss = date.getSeconds();
+
+    if (hh === 0) {
+      hh = 12;
+    }
+    if (hh > 12) {
+      hh = hh - 12;
+    }
+
+    hh = (hh < 10) ? "0" + hh : hh;
+    mm = (mm < 10) ? "0" + mm : mm;
+    ss = (ss < 10) ? "0" + ss : ss;
+
+    let time = hh + ":" + mm + ":" + ss;
+    document.getElementById("clock").innerText = time;
+  }
+  let t = setTimeout(function () { currentTime() }, 100);
+}
